Create paranthesis.java
Given an array of positive and negative integers, denoting different types of parentheses. 
The positive numbers  denotes opening parentheses of type  and negative number  denotes closing parentheses of type .
Open parentheses must be closed by the same type of parentheses. Open parentheses must be closed in the correct order, 
i.e., never close an open pair before its inner pair is closed (if it has an inner pair).
You have to find out the length of the longest subarray that is balanced.